requirements.txt
Django==5.1.6
django-phonenumber-field[phonenumbers]==8.0.0
python-dotenv==1.0.1
psycopg2==2.9.10
gunicorn==23.0.0
django-view-breadcrumbs==2.5.1
django-model-utils==5.0.0
whitenoise==6.9.0
pillow==11.1.0

mypy==1.15.0
django-stubs==5.1.3
djangorestframework-stubs==3.15.3
boto3-stubs==1.37.9

ruff==0.9.10
pre-commit==4.1.0
